move `useNwc2` into `useNwc`
