+++ src/pages/connect.tsx
@@ -10,7 +10,7 @@ import { useRouter } from 'next/router';
 import { NostrEvent } from '@nostr-dev-kit/ndk';
 import { useNDK } from '@/hooks/useNDK';
 import { addNwcConnection } from '@/redux/slices/NwcSlice';
-import { NWCMethods } from '@/hooks/useNwc2';
+import { NWCMethods } from '@/hooks/useNwc';
 
 export default function Home() {
    const router = useRouter();
